refactor: 도메인 폴더명 변경
